+++ wink-micro-os/tools/__init__.py
@@ -1,0 +1,1 @@
+# Makes ``tools`` importable when wink-micro-os is on PYTHONPATH.